Delete module & dto / move v3 to src / lint #375
