+++ backend/src/ban/ban.module.ts
@@ -1,6 +1,6 @@
 import { forwardRef, Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import BanLog from '../../entities/ban.log.entity';
+import BanLog from '../entities/ban.log.entity';
 import { UserModule } from '../user/user.module';
 import { BanService } from './ban.service';
 import { BanRepository } from './repository/ban.repository';
