+++ backend/src/blackhole/blackhole.module.ts
@@ -4,9 +4,9 @@ import { BlackholeService } from './blackhole.service';
 import { IBlackholeRepository } from './repository/blackhole.repository';
 import { RawqueryBlackholeRepository } from './repository/rawquery.blackhole.repository';
 import { AuthModule } from 'src/auth/auth.module';
-import { CabinetModule } from 'src/v3/cabinet/cabinet.module';
-import { LentModule } from 'src/v3/lent/lent.module';
-import { UserModule } from 'src/v3/user/user.module';
+import { CabinetModule } from 'src/cabinet/cabinet.module';
+import { LentModule } from 'src/lent/lent.module';
+import { UserModule } from 'src/user/user.module';
 
 const repo = {
   provide: IBlackholeRepository,
