+++ backend/src/event/event.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { AuthModule } from 'src/auth/auth.module';
-import { BanModule } from 'src/v3/ban/ban.module';
+import { BanModule } from 'src/ban/ban.module';
 import { EventController } from './event.controller';
 import { EventService } from './evnet.service';
 import { IEventRepository } from './repository/IEventRepository';
