+++ backend/src/utils/expired.checker.component.ts
@@ -1,16 +1,15 @@
-import {
-    forwardRef,
-  Inject,
-  Injectable,
-  Logger,
-} from '@nestjs/common';
+import { forwardRef, Inject, Injectable, Logger } from '@nestjs/common';
 import { Cron, CronExpression } from '@nestjs/schedule';
 import CabinetStatusType from 'src/enums/cabinet.status.type.enum';
 import { CabinetInfoService } from '../cabinet/cabinet.info.service';
 import { LentTools } from '../lent/lent.component';
 import { LentService } from '../lent/lent.service';
 import { EmailSender } from './email.sender.component';
-import { Transactional, Propagation, runOnTransactionComplete } from 'typeorm-transactional';
+import {
+  Transactional,
+  Propagation,
+  runOnTransactionComplete,
+} from 'typeorm-transactional';
 import Lent from 'src/entities/lent.entity';
 
 @Injectable()
@@ -40,24 +39,24 @@ export class ExpiredChecker {
   })
   async checkExpiredCabinetEach(lent: Lent) {
     const days = await this.getExpiredDays(lent.expire_time);
-      if (days >= 0) {
-        if (days > 0 && days < 15)
-          await this.cabinetInfoService.updateCabinetStatus(
-            lent.lent_cabinet_id,
-            CabinetStatusType.EXPIRED,
-          );
-        else if (days >= 15) {
-          await this.cabinetInfoService.updateCabinetStatus(
-            lent.lent_cabinet_id,
-            CabinetStatusType.BANNED,
-          );
-          await this.lentService.returnCabinet({
-            user_id: lent.lent_user_id,
-            intra_id: lent.user.intra_id,
-          });
-        }
-        this.emailsender.mailing(lent.user.intra_id, days);
+    if (days >= 0) {
+      if (days > 0 && days < 15)
+        await this.cabinetInfoService.updateCabinetStatus(
+          lent.lent_cabinet_id,
+          CabinetStatusType.EXPIRED,
+        );
+      else if (days >= 15) {
+        await this.cabinetInfoService.updateCabinetStatus(
+          lent.lent_cabinet_id,
+          CabinetStatusType.BANNED,
+        );
+        await this.lentService.returnCabinet({
+          user_id: lent.lent_user_id,
+          intra_id: lent.user.intra_id,
+        });
       }
+      this.emailsender.mailing(lent.user.intra_id, days);
+    }
     runOnTransactionComplete((err) => err && this.logger.error(err));
   }
 
@@ -68,8 +67,7 @@ export class ExpiredChecker {
     );
     const lentList = await Promise.all(await this.lentTools.getAllLent());
     lentList.forEach(async (lent: Lent) => {
-        await this.checkExpiredCabinetEach(lent);
-      }
-  );
+      await this.checkExpiredCabinetEach(lent);
+    });
   }
 }
