+++ dev/configure.conf
@@ -1,76 +1,10 @@
-map $http_upgrade $connection_upgrade {
-default upgrade;
-'' close;
-}
-
-map $remote_addr $proxy_forwarded_elem {
-    # IPv4 addresses can be sent as-is
-    ~^[0-9.]+$          "for=$remote_addr";
-
-    # IPv6 addresses need to be bracketed and quoted
-    ~^[0-9A-Fa-f:.]+$   "for=\"[$remote_addr]\"";
-
-    # Unix domain socket names cannot be represented in RFC 7239 syntax
-    default             "for=unknown";
-}
-
-map $http_forwarded $proxy_add_forwarded {
-    # If the incoming Forwarded header is syntactically valid, append to it
-    "~^(,[ \\t]*)*([!#$%&'*+.^_`|~0-9A-Za-z-]+=([!#$%&'*+.^_`|~0-9A-Za-z-]+|\"([\\t \\x21\\x23-\\x5B\\x5D-\\x7E\\x80-\\xFF]|\\\\[\\t \\x21-\\x7E\\x80-\\xFF])*\"))?(;([!#$%&'*+.^_`|~0-9A-Za-z-]+=([!#$%&'*+.^_`|~0-9A-Za-z-]+|\"([\\t \\x21\\x23-\\x5B\\x5D-\\x7E\\x80-\\xFF]|\\\\[\\t \\x21-\\x7E\\x80-\\xFF])*\"))?)*([ \\t]*,([ \\t]*([!#$%&'*+.^_`|~0-9A-Za-z-]+=([!#$%&'*+.^_`|~0-9A-Za-z-]+|\"([\\t \\x21\\x23-\\x5B\\x5D-\\x7E\\x80-\\xFF]|\\\\[\\t \\x21-\\x7E\\x80-\\xFF])*\"))?(;([!#$%&'*+.^_`|~0-9A-Za-z-]+=([!#$%&'*+.^_`|~0-9A-Za-z-]+|\"([\\t \\x21\\x23-\\x5B\\x5D-\\x7E\\x80-\\xFF]|\\\\[\\t \\x21-\\x7E\\x80-\\xFF])*\"))?)*)?)*$" "$http_forwarded, $proxy_forwarded_elem";
-
-    # Otherwise, replace it
-    default "$proxy_forwarded_elem";
-}
-
 server {
   listen       80  default_server;
   server_name  _;
 
-  # Vite
-  location /hmr/ {
-    proxy_pass              http://host.docker.internal:4242/;
-    proxy_http_version                 1.1;
-    proxy_cache_bypass                 $http_upgrade;
-
-    # Proxy headers
-    proxy_set_header Upgrade           $http_upgrade;
-    proxy_set_header Connection        $connection_upgrade;
-    proxy_set_header Host              $host;
-    proxy_set_header X-Real-IP         $remote_addr;
-    proxy_set_header Forwarded         $proxy_add_forwarded;
-    proxy_set_header X-Forwarded-For   $proxy_add_x_forwarded_for;
-    proxy_set_header X-Forwarded-Proto $scheme;
-    proxy_set_header X-Forwarded-Host  $host;
-    proxy_set_header X-Forwarded-Port  $server_port;
-    proxy_set_header Connection        $connection_upgrade;
-
-    # Proxy timeouts
-    proxy_connect_timeout              60s;
-    proxy_send_timeout                 3600s;
-    proxy_read_timeout                 3600s;
-  }
-
   location / {
     proxy_pass http://host.docker.internal:4242;
-    proxy_http_version                 1.1;
-    proxy_cache_bypass                 $http_upgrade;
-
-    # Proxy headers
-    proxy_set_header Upgrade           $http_upgrade;
-    proxy_set_header Connection        $connection_upgrade;
-    proxy_set_header Host              $host;
-    proxy_set_header X-Real-IP         $remote_addr;
-    proxy_set_header Forwarded         $proxy_add_forwarded;
-    proxy_set_header X-Forwarded-For   $proxy_add_x_forwarded_for;
-    proxy_set_header X-Forwarded-Proto $scheme;
-    proxy_set_header X-Forwarded-Host  $host;
-    proxy_set_header X-Forwarded-Port  $server_port;
-    proxy_set_header Connection        $connection_upgrade;
-
-    # Proxy timeouts
-    proxy_connect_timeout              60s;
-    proxy_send_timeout                 3600s;
-    proxy_read_timeout                 3600s;
+    proxy_set_header Host $host;
   }
 
   location ^~ /api {
@@ -79,11 +13,6 @@ server {
   }
 
   location ^~ /auth {
-    proxy_pass http://host.docker.internal:2424;
-    proxy_set_header Host $host;
-  }
-
-  location ^~ /v3/api {
     proxy_pass http://host.docker.internal:2424;
     proxy_set_header Host $host;
   }
